--- lib/src/utils/setup/check_dependencies.rs
@@ -1,20 +1,0 @@
-use crate::{check_cabal, check_ghc, check_ghcup, check_libsodium, print};
-use anyhow::Result;
-
-pub async fn check_dependencies() -> Result<()> {
-    print("", "Checking dependencies")?;
-    check_ghcup().await?;
-    check_ghc().await?;
-    check_cabal().await?;
-    check_libsodium().await
-}
-
-#[cfg(test)]
-mod test {
-    // use crate::check_dependencies;
-    #[tokio::test]
-    #[ignore]
-    async fn test_check_dependencies() {
-        unimplemented!();
-    }
-}
