--- lib/src/utils/setup/check_install.rs
@@ -1,25 +1,0 @@
-use crate::{check_installed_version, print, print_emoji, source_shell};
-use anyhow::Result;
-use console::Emoji;
-
-pub async fn check_install(component: &str) -> Result<()> {
-    let msg = format!("Checking successful {} installation", component);
-    print("", &msg)?;
-    if let "cardano-node" = component {
-        check_installed_version("cardano-cli").await?;
-    }
-    check_installed_version(component).await?;
-    source_shell().await?;
-    let msg = format!("Successfully installed {}", component);
-    print_emoji("green", &msg, Emoji("🙌🎉", ""))
-}
-
-#[cfg(test)]
-mod test {
-    // use crate::check_install;
-    #[tokio::test]
-    #[ignore]
-    async fn test_check_install() {
-        unimplemented!();
-    }
-}
