--- lib/src/utils/setup/get_component_path.rs
@@ -1,20 +1,0 @@
-use crate::check_env;
-use anyhow::Result;
-use convert_case::{Case, Casing};
-
-pub async fn get_component_path(component: &str) -> Result<String> {
-    let env = format!("{}-dir", component);
-    let converted = env.to_case(Case::UpperSnake);
-    let path = check_env(&converted)?;
-    Ok(path)
-}
-
-#[cfg(test)]
-mod test {
-    // use crate::get_component_path;
-    #[tokio::test]
-    #[ignore]
-    async fn test_get_component_path() {
-        unimplemented!();
-    }
-}
