--- lib/src/utils/setup/mod.rs
@@ -1,20 +1,0 @@
-pub mod prepare_build;
-pub use prepare_build::prepare_build;
-pub mod configure_build;
-pub use configure_build::configure_build;
-pub mod update_project_file;
-pub use update_project_file::update_project_file;
-pub mod check_dependencies;
-pub use check_dependencies::check_dependencies;
-pub mod install_component;
-pub use install_component::install_component;
-pub mod check_install;
-pub use check_install::check_install;
-pub mod build_component;
-pub use build_component::build_component;
-pub mod get_component_path;
-pub use get_component_path::get_component_path;
-pub mod get_project_file;
-pub use get_project_file::get_project_file;
-pub mod set_confirm;
-pub use set_confirm::set_confirm;
