--- lib/src/utils/setup/set_confirm.rs
@@ -1,9 +1,0 @@
-use crate::set_env;
-
-pub fn set_confirm(confirm: bool) {
-    if confirm {
-        set_env("CONFIRM", "true")
-    } else {
-        set_env("CONFIRM", "false")
-    }
-}
